+++ src/main/java/cl/bozz/nqueensdfs/utils/TotalPermutationsCalculator.kt
@@ -1,10 +1,10 @@
 package cl.bozz.nqueensdfs.utils
 
 object TotalPermutationsCalculator {
-    fun totalPermutations(boardSize: Int): Long {
+    fun totalPermutations(boardSize: UInt): ULong {
         val totalCells = boardSize * boardSize
-        var result = totalCells.toLong()
-        for (i in 1 until boardSize) {
+        var result = totalCells.toULong()
+        for (i in 1.toUInt() until boardSize) {
             val mult = totalCells - i
             result *= mult
         }
